Split extension on different files
