+++ TagsCloudVisualization/TagsCloudVisualization/Helpers/PointExtensions.cs
@@ -1,0 +1,16 @@
+﻿using System.Drawing;
+
+namespace TagsCloudVisualization.Helpers;
+
+public static class PointExtensions
+{
+	public static Point Plus(this Point point, Point otherPoint)
+	{
+		return new Point(point.X + otherPoint.X, point.Y + otherPoint.Y);
+	}
+
+	public static Point Minus(this Point point, Point otherPoint)
+	{
+		return new Point(point.X - otherPoint.X, point.Y - otherPoint.Y);
+	}
+}
